Implement recording functionality

+++ src/gather/writer.rs
@@ -1,12 +1,15 @@
 #[cfg(feature = "archive")]
 use flate2::{write::GzEncoder, Compression};
 
+use json_patch::diff;
+use k8s_openapi::{chrono::Utc, serde_json};
 use serde::Deserialize;
 use std::{
     ffi::OsStr,
     fmt::Display,
     fs::{DirBuilder, File},
     io::Write as _,
+    ops::Deref,
     path::PathBuf,
     sync::{Arc, Mutex},
 };
@@ -15,6 +18,8 @@ use tar::{Builder, Header};
 use walkdir::WalkDir;
 #[cfg(feature = "archive")]
 use zip::{write::SimpleFileOptions, ZipWriter};
+
+use crate::gather::reader::Reader;
 
 use super::representation::{ArchivePath, Representation};
 
@@ -198,11 +203,13 @@ impl Writer {
         match self {
             Self::Path(Archive(archive)) => {
                 let file = archive.join(archive_path);
-                DirBuilder::new()
-                    .recursive(true)
-                    .create(file.parent().unwrap())?;
-                let mut file = File::create(file)?;
-                file.write_all(data.as_bytes())?;
+                if !file.exists() {
+                    DirBuilder::new()
+                        .recursive(true)
+                        .create(file.parent().unwrap())?;
+                    let mut file = File::create(file)?;
+                    file.write_all(data.as_bytes())?;
+                }
             }
             #[cfg(feature = "archive")]
             Self::Gzip(Archive(archive), builder) => {
@@ -226,6 +233,42 @@ impl Writer {
                 let file = file.to_str().unwrap();
                 writer.start_file(file, SimpleFileOptions::default())?;
                 writer.write_all(data.as_bytes())?;
+            }
+        }
+        Ok(())
+    }
+
+    /// Adds a representation data to the archive under the representation path
+    pub fn sync(&mut self, repr: &Representation) -> anyhow::Result<()> {
+        log::debug!("Writing {}...", repr.path());
+
+        let archive_path: String = repr.path().try_into()?;
+
+        match self {
+            Self::Path(archive) => {
+                let file_path = archive.0.join(archive_path);
+
+                // generate diff and write
+                let original = Reader::new(archive.clone(), Utc::now())?.read(file_path.clone())?;
+                let updated = serde_yaml::from_str(repr.data())?;
+                let patch = &diff(&original, &updated);
+                if !patch.deref().is_empty() {
+                    let mut patches = File::options()
+                        .create(true)
+                        .append(true)
+                        .open(file_path.with_extension("patch"))?;
+                    serde_json::to_writer(patches.try_clone()?, patch)?;
+                    patches.write_all(b"\n")?;
+                }
+                self.store(repr)?;
+            }
+            #[cfg(feature = "archive")]
+            Self::Gzip(Archive(_archive), _builder) => {
+                unimplemented!();
+            }
+            #[cfg(feature = "archive")]
+            Self::Zip(Archive(_archive), _writer) => {
+                unimplemented!();
             }
         }
         Ok(())
@@ -373,7 +416,7 @@ mod tests {
     fn test_try_into_nested_file_success() {
         let tmp_dir = TempDir::new("archive").expect("failed to create temp dir");
         let tmp_dir = tmp_dir.path();
-        let mut writer = Writer::new(
+        let writer = Writer::new(
             &Archive::new(tmp_dir.join("nested/output.zip")),
             &Encoding::Zip,
         )
